requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Login>=0.6,<1.0
Flask-WTF>=1.2,<2.0
Flask-Migrate>=4.0,<5.0
Flask-Mail>=0.10,<1.0
Flask-SocketIO>=5.5,<6.0
python-dotenv>=1.0,<2.0
email-validator>=2.0,<3.0
PyMySQL>=1.1,<2.0
bcrypt>=4.1,<5.0
requests-oauthlib>=2.0,<3.0
